+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### 3.0 (WIP)
+
+  * Minimal PHP version required bumped to PHP7
+
+#### Instantiator
+
+  * Introduced a new interface `InstantiatorInterface`
+  * Changed instantiator methods to be chainable instantiators instead
+
 ### 2.1.2 (2015-12-10)
 
   * Bug fix - private properties were populatable directly on the instance, but not private properties of parent classes. Although this is an antipattern, if we're allowing it for the instance we should allow it up the chain.